refactor: extract kebab checks from `check_import_map.ts` to a new file
